refactor, add ReadData.materialize()
This is in preparation for moving SplittableReadData into a separate PR
