+++ src/main/java/org/janelia/saalfeldlab/n5/readdata/AbstractInputStreamReadData.java
@@ -11,6 +11,11 @@ abstract class AbstractInputStreamReadData implements ReadData {
 
 	@Override
 	public SplittableReadData splittable() throws IOException {
+		return materialize();
+	}
+
+	@Override
+	public SplittableReadData materialize() throws IOException {
 		if (bytes == null) {
 			final byte[] data;
 			final int length = (int) length();
@@ -27,6 +32,6 @@ abstract class AbstractInputStreamReadData implements ReadData {
 
 	@Override
 	public byte[] allBytes() throws IOException, IllegalStateException {
-		return splittable().allBytes();
+		return materialize().allBytes();
 	}
 }
