+++ src/main/java/org/janelia/saalfeldlab/n5/readdata/LazyReadData.java
@@ -4,11 +4,29 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import org.apache.commons.io.output.ProxyOutputStream;
 
 class LazyReadData implements ReadData {
 
 	LazyReadData(final OutputStreamWriter writer) {
 		this.writer = writer;
+	}
+
+	/**
+	 * Construct a {@code LazyReadData} that uses the given {@code OutputStreamEncoder} to
+	 * encode the given {@code ReadData}.
+	 *
+	 * @param data
+	 * 		the ReadData to encode
+	 * @param encoder
+	 * 		OutputStreamEncoder to use for encoding
+	 */
+	LazyReadData(final ReadData data, final OutputStreamOperator encoder) {
+		this(outputStream -> {
+			try (final OutputStream deflater = encoder.apply(interceptClose.apply(outputStream))) {
+				data.writeTo(deflater);
+			}
+		});
 	}
 
 	private final OutputStreamWriter writer;
@@ -17,6 +35,11 @@ class LazyReadData implements ReadData {
 
 	@Override
 	public SplittableReadData splittable() throws IOException {
+		return materialize();
+	}
+
+	@Override
+	public SplittableReadData materialize() throws IOException {
 		if (bytes == null) {
 			final ByteArrayOutputStream baos = new ByteArrayOutputStream(8192);
 			writeTo(baos);
@@ -27,17 +50,17 @@ class LazyReadData implements ReadData {
 
 	@Override
 	public long length() throws IOException {
-		return splittable().length();
+		return materialize().length();
 	}
 
 	@Override
 	public InputStream inputStream() throws IOException, IllegalStateException {
-		return splittable().inputStream();
+		return materialize().inputStream();
 	}
 
 	@Override
 	public byte[] allBytes() throws IOException, IllegalStateException {
-		return splittable().allBytes();
+		return materialize().allBytes();
 	}
 
 	@Override
@@ -48,4 +71,16 @@ class LazyReadData implements ReadData {
 			writer.writeTo(outputStream);
 		}
 	}
+
+	/**
+	 * {@code UnaryOperator} that wraps {@code OutputStream} to intercept {@code
+	 * close()} and call {@code flush()} instead
+	 */
+	private static OutputStreamOperator interceptClose = o -> new ProxyOutputStream(o) {
+
+		@Override
+		public void close() throws IOException {
+			out.flush();
+		}
+	};
 }
