+++ src/main/java/org/janelia/saalfeldlab/n5/readdata/ReadData.java
@@ -5,7 +5,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import org.janelia.saalfeldlab.n5.KeyValueAccess;
-import org.janelia.saalfeldlab.n5.readdata.EncodedReadData.OutputStreamOperator;
 
 /**
  * An abstraction over {@code byte[]} data.
@@ -99,6 +98,16 @@ public interface ReadData {
 	}
 
 	/**
+	 * Read the underlying data into a {@code byte[]} array, and return it as a {@code ReadData}.
+	 * (If this {@code ReadData} is already in a {@code byte[]} array or {@code
+	 * ByteBuffer}, just return {@code this}.)
+	 * <p>
+	 * The returned {@code ReadData} has a known {@link #length} and multiple
+	 * {@link #inputStream InputStreams} can be opened on it.
+	 */
+	SplittableReadData materialize() throws IOException;
+
+	/**
 	 * If this {@code ReadData} is a {@code SplittableReadData}, just returns {@code this}.
 	 * <p>
 	 * Otherwise, if the underlying data is an {@code InputStream}, all data is read and
@@ -144,7 +153,17 @@ public interface ReadData {
 	 * @return encoded ReadData
 	 */
 	default ReadData encode(OutputStreamOperator encoder) {
-		return new EncodedReadData(this, encoder);
+		return new LazyReadData(this, encoder);
+	}
+
+	/**
+	 * Like {@code UnaryOperator<OutputStream>}, but {@code apply} throws {@code IOException}.
+	 */
+	@FunctionalInterface
+	interface OutputStreamOperator {
+
+		OutputStream apply(OutputStream o) throws IOException;
+
 	}
 
 	// --------------- Factory Methods ------------------
@@ -247,6 +266,14 @@ public interface ReadData {
 		void writeTo(OutputStream outputStream) throws IOException, IllegalStateException;
 	}
 
+	/**
+	 * Create a new {@code ReadData} that is lazily generated by the given {@link OutputStreamWriter}.
+	 *
+	 * @param generator
+	 * 		generates the data
+	 *
+	 * @return a new ReadData
+	 */
 	static ReadData from(OutputStreamWriter generator) {
 		return new LazyReadData(generator);
 	}
